Racket: fix for module evaluation/entering
Our module loader is receiving load requests for module names
represented as lists that are not exactly a submodule, in the sense
that the path does not represent an actual file.

This phenomenon happens for instance when specifying a reader in a
#lang tag.  E.g.

   #lang at-exp racket

will cause the loader to be called with module name '(main reader) and
path <cols-path>/at-exp/main.rkt, where main.rkt does not exist.
Afterwards, we see a call to load at-exp/lang/reader/rkt, with name
reader, which is the real code.

So, for now, i'm skipping all load requests with a list name,
forwarding them to racket's default loader.