racket: struggling with submodules
Submodule (re)loading is not without pecularities.  In particular,
module[*+] submodules are not visited the first time one enters its
parent, but once you load them once, they're revisited every time we
load the parent afterwards--racket's native enter! exhibits the same
behaviour, so i'm guessing we'll have to live with that.

There is however a glitch in that submodules can only be reloaded then
by loading the parent, so we need to confirm that this is expected
behaviour and, if it is, automating the parent's load when the
submodule's is requested.

On the other hand, entering a module[*+] is not working in Geiser yet,
and it does in plain racket, so this one is our fault.  Working on it.