Avoiding a segfault due to out of bounds stack size
The parser in geiser-syntax is (tail, but elisp doesn't care)
recursive, and we are setting max-lisp-eval-depth to some, ahem,
heuristic value before starting a read.  For long strings, such as
that returned by the list of identifiers exported by the racket
module, the heuristic was bad enough to produce a value making Emacs
to blow away.

This is just a palliative.  The real solution is turn the recursion in
geiser-syntax--read into an explicit iteration.