Paving the way to MELPA (and simpler ELPA generation)
Following the discussion in GitHub, i'm adding an alternative scheme
path to geiser.el as well as ##autoload cookies.  As a first benefit,
this simplifies a bit the elpa target, after playing a similar path
trick in bin/geiser-racket.sh.

Things should be almost ready for creating a MELPA recipe.