Using font-lock-variable-name-face in lieu of bold for autodoc
For some X faces, a bold string in the modeline causes emacs to widen it
to two lines, which is kind of annoying.  The default value of
font-lock-variable-name-face on color/X displays doesn't include any
boldness, and will probably improve the default experience of new users.
Thanks to Mario Domenech Goulart for noticing this and the previous one!