New commands geiser-repl-{previous,next}-prompt
When using a prompt regexp, comint's version of these commands
misbehave (they try to reuse forward-paragraph, and that's not quite
it), so we're implemeting our own here in a very straightforward way.
We also bind the usual C-c C-p and C-c C-n to them.  It only remains to
b seen whether advising the original ones is worth the trouble.