Not killing partial outputs when leaving the REPL
Since this job is done in the process sentinel, the clean up is also
triggered when the Scheme process exits unexpectedly, deleting any
traces the dying guy might have left.  I added a flag to control the
behaviour, but upon reflection the old behaviour seems wrong and i've
defaulted to the new one.  This one should fix #251.